Views refactor #1 - extract out edit judgment view
This begins the process of extracting the views out of the mammoth views.py file, starting with the Edit Judgment view. I've started moving non-view-related concerns out into utils too, further commits will tidy that up a bit :-)
