+++ judgments/utils.py
@@ -1,4 +1,7 @@
+import json
 import logging
+import re
+import time
 import xml.etree.ElementTree as ET
 from datetime import datetime
 
@@ -11,8 +14,12 @@ from caselawclient.Client import (
     MarklogicResourceNotFoundError,
     api_client,
 )
+from django.contrib.auth.models import User
 
 env = environ.Env()
+
+
+VERSION_REGEX = r"(\d+)-(\d+|TDR)"
 
 
 class MoveJudgmentError(Exception):
@@ -130,18 +137,182 @@ def build_new_key(old_key, new_uri):
         return f"{new_uri}/{old_filename}"
 
 
-def create_s3_client():
+def create_aws_client(service: str):  # service
+    """@param: service The AWS service, e.g. 's3'"""
     aws = boto3.session.Session(
         aws_access_key_id=env("AWS_ACCESS_KEY_ID", default=None),
         aws_secret_access_key=env("AWS_SECRET_KEY", default=None),
     )
     return aws.client(
-        "s3",
+        service,
         endpoint_url=env("AWS_ENDPOINT_URL", default=None),
         region_name=env("PRIVATE_ASSET_BUCKET_REGION", default=None),
         config=botocore.client.Config(signature_version="s3v4"),
     )
 
 
+def create_s3_client():
+    return create_aws_client("s3")
+
+
 def uri_for_s3(uri: str):
     return uri.lstrip("/")
+
+
+def generate_docx_url(uri: str):
+    # If there isn't a PRIVATE_ASSET_BUCKET, don't try to get the bucket.
+    # This helps local environment setup where we don't use S3.
+    bucket = env("PRIVATE_ASSET_BUCKET", None)
+    if not bucket:
+        return ""
+
+    client = create_s3_client()
+
+    key = f'{uri}/{uri.replace("/", "_")}.docx'
+
+    return client.generate_presigned_url(
+        "get_object", Params={"Bucket": bucket, "Key": key}
+    )
+
+
+def generate_pdf_url(uri: str):
+    bucket = env("PRIVATE_ASSET_BUCKET", None)
+    if not bucket:
+        return ""
+
+    client = create_s3_client()
+
+    key = f'{uri}/{uri.replace("/", "_")}.pdf'
+
+    return client.generate_presigned_url(
+        "get_object", Params={"Bucket": bucket, "Key": key}
+    )
+
+
+def delete_from_bucket(uri: str, bucket: str) -> None:
+    client = create_s3_client()
+    response = client.list_objects(Bucket=bucket, Prefix=uri)
+
+    if response.get("Contents"):
+        objects_to_delete = [
+            {"Key": obj["Key"]} for obj in response.get("Contents", [])
+        ]
+        client.delete_objects(
+            Bucket=bucket,
+            Delete={
+                "Objects": objects_to_delete,
+            },
+        )
+
+
+def delete_documents(uri: str) -> None:
+    unpublish_documents(uri)
+    delete_from_bucket(uri, env("PRIVATE_ASSET_BUCKET"))
+
+
+def publish_documents(uri: str) -> None:
+    client = create_s3_client()
+
+    public_bucket = env("PUBLIC_ASSET_BUCKET")
+    private_bucket = env("PRIVATE_ASSET_BUCKET")
+
+    response = client.list_objects(Bucket=private_bucket, Prefix=uri)
+
+    for result in response.get("Contents", []):
+        key = str(result["Key"])
+
+        if not key.endswith("parser.log") and not key.endswith(".tar.gz"):
+            source = {"Bucket": private_bucket, "Key": key}
+            extra_args = {"ACL": "public-read"}
+            try:
+                client.copy(source, public_bucket, key, extra_args)
+            except botocore.client.ClientError as e:
+                logging.warning(
+                    f"Unable to copy file {key} to new location {public_bucket}, error: {e}"
+                )
+
+
+def unpublish_documents(uri: str) -> None:
+    delete_from_bucket(uri, env("PUBLIC_ASSET_BUCKET"))
+
+
+def notify_changed(uri: str, status: str, enrich: bool = False) -> None:
+    client = create_aws_client("sns")
+
+    message_attributes = {}
+    message_attributes["update_type"] = {
+        "DataType": "String",
+        "StringValue": status,
+    }
+    message_attributes["uri_reference"] = {
+        "DataType": "String",
+        "StringValue": uri,
+    }
+    if enrich:
+        message_attributes["trigger_enrichment"] = {
+            "DataType": "String",
+            "StringValue": "1",
+        }
+
+    client.publish(
+        TopicArn=env("SNS_TOPIC"),
+        Message=json.dumps({"uri_reference": uri, "status": status}),
+        Subject=f"Updated: {uri} {status}",
+        MessageAttributes=message_attributes,
+    )
+
+
+def invalidate_caches(uri: str) -> None:
+    if (
+        env("CLOUDFRONT_INVALIDATION_ACCESS_KEY_ID", default=None) is None
+        and env("CLOUDFRONT_INVALIDATION_ACCESS_SECRET", default=None) is None
+    ):
+        logging.warning(
+            "Cannot invalidate cache: no cloudfront environment variables set"
+        )
+        return
+
+    aws = boto3.session.Session(
+        aws_access_key_id=env("CLOUDFRONT_INVALIDATION_ACCESS_KEY_ID", default=None),
+        aws_secret_access_key=env(
+            "CLOUDFRONT_INVALIDATION_ACCESS_SECRET", default=None
+        ),
+    )
+    cloudfront = aws.client("cloudfront")
+    cloudfront.create_invalidation(
+        DistributionId=env("CLOUDFRONT_PUBLIC_DISTRIBUTION_ID"),
+        InvalidationBatch={
+            "Paths": {"Quantity": 1, "Items": ["/*"]},
+            "CallerReference": str(time.time()),
+        },
+    )
+
+
+def render_versions(decoded_versions):
+    versions = [
+        {
+            "uri": part.text.rstrip(".xml"),
+            "version": extract_version(part.text),
+        }
+        for part in decoded_versions
+    ]
+    sorted_versions = sorted(versions, key=lambda d: -d["version"])
+    return sorted_versions
+
+
+def extract_version(version_string: str) -> int:
+    try:
+        return int(re.search(VERSION_REGEX, version_string).group(1))
+    except AttributeError:
+        return 0
+
+
+def users_dict():
+    users = User.objects.all()
+    return [
+        {
+            "name": u.get_username(),
+            "print_name": u.get_full_name() or u.get_username(),
+        }
+        for u in users
+    ]
